+++ src/kiara/rendering/__init__.py
@@ -2,8 +2,8 @@
 import typing
 
 from kiara import Kiara, PipelineController
-from kiara.config import KiaraWorkflowConfig
 from kiara.data import ValueSet
+from kiara.module_config import KiaraWorkflowConfig
 from kiara.workflow import KiaraWorkflow